backend/requirements.txt
Flask
Flask-CORS
python-dotenv
openai
moviepy
pydub
flask_cors